Remove `creates` property for `drush init`
Under Drush `9.x`, the initialization was failing in the case of
a pre-provisioned box which didn't yet have `drush.yml`.

That caused aliases to not work.

`drush init` is inherently idempotent.  It won't copy files that
already exist, and it won't add to `.bash_profile` if the additions
are already there.